auth,client/auth: move client related data structures to client package
