+++ auth.go
@@ -14,40 +14,13 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
 	"os"
 	"strings"
 
 	"github.com/mendersoftware/log"
+	"github.com/mendersoftware/mender/client"
 	"github.com/pkg/errors"
 )
-
-type AuthToken string
-
-type AuthReqData struct {
-	IdData      string `json:"id_data"`
-	TenantToken string `json:"tenant_token"`
-	Pubkey      string `json:"pubkey"`
-	SeqNumber   uint64 `json:"seq_no"`
-}
-
-type AuthRequest struct {
-	// request message data
-	Data []byte
-	// authorization token
-	Token AuthToken
-	// request signature
-	Signature []byte
-}
-
-// handler for authorization message data
-type AuthDataMessenger interface {
-	// build authorization request data, returns auth request or an error
-	MakeAuthRequest() (*AuthRequest, error)
-	// receive authoiriation data response, returns error if response is invalid
-	RecvAuthResponse([]byte) error
-}
 
 type AuthManager interface {
 	// returns true if authorization data is current and valid
@@ -61,14 +34,14 @@ type AuthManager interface {
 	// generate device key (will overwrite an already existing key)
 	GenerateKey() error
 
-	AuthDataMessenger
+	client.AuthDataMessenger
 }
 
 const (
 	authTokenName = "authtoken"
 	authSeqName   = "authseq"
 
-	noAuthToken = AuthToken("")
+	noAuthToken = client.EmptyToken
 )
 
 type MenderAuthManager struct {
@@ -76,7 +49,7 @@ type MenderAuthManager struct {
 	keyStore    *Keystore
 	idSrc       IdentityDataGetter
 	seqNum      SeqnumGetter
-	tenantToken AuthToken
+	tenantToken client.AuthToken
 }
 
 type AuthManagerConfig struct {
@@ -98,7 +71,7 @@ func NewAuthManager(conf AuthManagerConfig) AuthManager {
 		keyStore:    conf.KeyStore,
 		idSrc:       conf.IdentitySource,
 		seqNum:      NewFileSeqnum(authSeqName, conf.AuthDataStore),
-		tenantToken: AuthToken(conf.TenantToken),
+		tenantToken: client.AuthToken(conf.TenantToken),
 	}
 
 	if err := mgr.keyStore.Load(); err != nil && !IsNoKeys(err) {
@@ -158,15 +131,7 @@ func (m *MenderAuthManager) MakeAuthRequest() (*AuthRequest, error) {
 
 	log.Debugf("authorization data: %v", authd)
 
-	databuf := &bytes.Buffer{}
-	enc := json.NewEncoder(databuf)
-
-	err = enc.Encode(&authd)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to encode auth request")
-	}
-
-	reqdata := databuf.Bytes()
+	reqdata := authd.ToBytes()
 
 	// generate signature
 	sig, err := m.keyStore.Sign(reqdata)
@@ -176,7 +141,7 @@ func (m *MenderAuthManager) MakeAuthRequest() (*AuthRequest, error) {
 
 	return &AuthRequest{
 		Data:      reqdata,
-		Token:     AuthToken(tentok),
+		Token:     client.Token(tentok),
 		Signature: sig,
 	}, nil
 }
